Refactors notification utils to make them testable Adds sendRoundNotifications on notificationUtils.js
